tests: add basic geoconvert test and move tests to test_geoconvert.rs
